+++ odoo/local-src/connector_esb/models/customer_address/exporter.py
@@ -41,7 +41,7 @@ class CustomerAddressExportMapper(Component):
         if record.parent_id:
             return {'CustomerId': record.parent_id.ref}
         else:
-            return {'CustomerId': ''}
+            return {'CustomerId': record.ref}
 
     @mapping
     def compute_street(self, record):
@@ -59,19 +59,11 @@ class CustomerAddressExportMapper(Component):
 
     @mapping
     def compute_isdefaults(self, record):
-        """ If there is only one address for a customer, either invoice or delivery
-            then it is also the default for the other type of address
-        """
-        # Find out if the other type of address exist
-        other_type = 'invoice' if record.type == 'delivery' else 'delivery'
-        other_exists = self.env['res.partner'].search_count([
-            ('parent_id', '=', record.parent_id.id),
-            ('type', '=', other_type)])
+        """Set the type of address being created it is one or the other"""
+        is_invoicing_address = self.options.address_kind == 'invoice'
         return {
-            'IsDefaultBilling': ((record.type == 'invoice')
-                                 or (other_exists == 0)),
-            'IsDefaultShipping': ((record.type == 'delivery')
-                                  or (other_exists == 0))
+            'IsDefaultBilling': (is_invoicing_address),
+            'IsDefaultShipping': (not is_invoicing_address)
         }
 
 
@@ -87,5 +79,71 @@ class CustomerAddressCronExporter(Component):
         return bool(work.timestamp and
                     work.timestamp.kind == 'customer.address')
 
+    def _prepare_item(self, items):
+        prepared = []
+        for kind, item in items:
+            prepared.append(
+                self.mapper.map_record(item).values(address_kind=kind))
+        return prepared
+
+    def get_items(self, export_since):
+        """Get customer addresses and add type of address to export
+
+        Export for each customer the invoice and delivery address
+        And if the specific address does not exist the default address
+        of the customer must be used.
+        For the mapper to know which address is exported, the type is included
+        with each item.
+        """
+        items = super(CustomerAddressCronExporter,
+                      self).get_items(export_since)
+        # get_items will return all modified partners including the addresses.
+        # Wwe need to get the related parent or children to be sure we include
+        # the pair (invoice, delivery), even if they have not been modified.
+        # The following search extend the items with the children addresses
+        # or the parent.
+        items = self.env['res.partner'].search(
+                ['|',
+                 ('id', 'child_of', items.ids),
+                 ('child_ids', 'in', items.ids),
+                 ('type', 'in', ('delivery', 'invoice', 'contact')),
+                 ],
+                order='create_date DESC'
+                )
+        # group the records by kind of address in dictionaries for fast lookups
+        customers = []
+        invoice_addresses = {}
+        delivery_addresses = {}
+        for item in items:
+            parent_id = item.parent_id.id
+            if item.type == 'invoice':
+                if not invoice_addresses.get(parent_id):
+                    invoice_addresses[parent_id] = item
+            elif item.type == 'delivery':
+                if not delivery_addresses.get(parent_id):
+                    delivery_addresses[parent_id] = item
+            elif item.commercial_partner_id == item:
+                # Ignore items of type 'contact' which are children of a
+                # partner. E.g. prevent to export 'City Z' as both invoice
+                # and delivery for itself in this scenario:
+                #   John Doe (contact)
+                #   - City X (delivery)
+                #   - City Y (invoice)
+                #   - City Z (contact)
+                customers.append(item)
+
+        # Get the address to export for each kind. We can loop on customers as
+        # we did a search for the parent partner previously. Even if only the
+        # invoice address has been modified, the parent partner should be
+        # in the list.
+        items2export = []
+        for customer in customers:
+            invoice = invoice_addresses.get(customer.id) or customer
+            items2export.append(('invoice', invoice))
+            delivery = delivery_addresses.get(customer.id) or customer
+            items2export.append(('delivery', delivery))
+
+        return items2export
+
     def get_items_domain(self):
-        return [('type', 'in', ['delivery', 'invoice'])]
+        return [('customer', '=', 1)]